carousel fixes after feedback
