+++ src/app/page.tsx
@@ -16,8 +16,11 @@ import { detectBrowserLanguage } from "./utils";
 import { i18nSlice } from "./state/i18nSlice";
 
 import Cookies from "js-cookie";
-import { PromoBanner, PromoBannerProps } from "components/PromoBanner";
-import Carousel from "components/Carousel";
+
+import {
+  PromoBannerCarousel,
+  PromoBannerProps,
+} from "components/PromoBannerCarousel";
 
 // Configuration for promo banner
 // Once both images and a targetUrl are defined the banner will automatically show
@@ -58,17 +61,7 @@ export default function Home() {
 
   return (
     <>
-      <Carousel
-        items={[
-          {
-            Content: () => (
-              <div className="flex justify-center">
-                <PromoBanner {...promoBannerConfig} />
-              </div>
-            ),
-          },
-        ]}
-      />
+      <PromoBannerCarousel items={[promoBannerConfig]} />
       <div className="flex-grow grid grid-cols-12 custom-auto-row-grids max-w-none divide-y-4 divide-base-300">
         <div className="col-span-12 lg:col-span-5 xl:col-span-3 text-center lg:border-r-4 border-t-4 border-base-300">
           <PairSelector />
